# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 拼单平台接口
  version: 0.2.0
  description: |
    T001：框架元数据与健康检查。
    T002：后台身份权限、商品图片、商品库存管理、小程序商品展示。
    金额一律整数分；数量为十进制字符串（≤3 位小数）；时间 ISO 8601 UTC；
    分页 page 从 1 起，pageSize 默认 10、上限 50；响应包络 {data, requestId}，错误 {code, message, requestId, details?}。
servers:
  - url: http://127.0.0.1:3000
    description: 本地联调
paths:
...
  /api/admin/v1/fulfillment/groups:
    get:
      operationId: listFulfillmentGroups
      security: [{ bearerAuth: [] }]
      summary: 履约组分页（order:manage；status=组内履约单状态汇总口径：全部同一状态才归入）
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending_shipment, partially_shipped, shipped, completed] } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 10 } }
      responses:
        '200':
          description: 分页履约组列表（各状态计数）
          content:
            application/json:
              schema:
                type: object
                required: [data, requestId]
                properties:
                  requestId: { type: string }
                  data:
                    type: object
                    required: [items, total]
                    properties:
                      items:
                        type: array
                        items:
                          type: object
                          required: [groupId, productId, productName, succeededAt, total, pending, partially, shipped, completed]
                          properties:
                            groupId: { type: string, format: uuid }
                            productId: { type: string, format: uuid }
                            productName: { type: string }
                            succeededAt: { type: string, format: date-time }
                            total: { type: integer }
                            pending: { type: integer }
                            partially: { type: integer }
                            shipped: { type: integer }
                            completed: { type: integer }
                      total: { type: integer }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
  /api/admin/v1/fulfillment/groups/{id}:
    get:
      operationId: getFulfillmentGroup
      security: [{ bearerAuth: [] }]
      summary: 组内履约明细（每用户分配数量/收货脱敏/包裹运单；order:manage）
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: 组内履约明细
          content:
            application/json:
              schema:
                type: object
                required: [data, requestId]
                properties:
                  requestId: { type: string }
                  data:
                    type: object
                    required: [groupId, items]
                    properties:
                      groupId: { type: string, format: uuid }
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/AdminFulfillmentDetail' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '404': { $ref: '#/components/responses/Error404' }
  /api/admin/v1/fulfillment/groups/{id}/shipments/export:
    get:
      operationId: exportFulfillmentShipments
      security: [{ bearerAuth: [] }]
      summary: 发货单导出 CSV（含收货明文；写审计 fulfillment.export；order:manage）
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: CSV（UTF-8 BOM）
          content:
            text/csv:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '404': { $ref: '#/components/responses/Error404' }
  /api/admin/v1/fulfillment/orders/{id}/shipments:
    post:
      operationId: shipFulfillmentOrder
      security: [{ bearerAuth: [] }]
      summary: 正常发货（最小履约单位；数量守恒；isReissue:true 返回409 REVIEW_REQUIRED）
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantityGrams, company, trackingNo]
              properties:
                quantityGrams: { type: integer, minimum: 1, description: 本次寄发数量（最小单位，重量为克、计数为商品单位） }
                company: { type: string, maxLength: 30 }
                trackingNo: { type: string, maxLength: 64 }
                isReissue: { type: boolean, default: false, description: 兼容字段；true返回409 REVIEW_REQUIRED，补发请使用售后审批 }
                reason: { type: string, maxLength: 200, description: 补发原因（isReissue 必填） }
      responses:
        '201':
          description: 已发货
          content:
            application/json:
              schema:
                type: object
                required: [data, requestId]
                properties:
                  requestId: { type: string }
                  data:
                    type: object
                    required: [shipmentId, fulfillmentOrder]
                    properties:
                      shipmentId: { type: string, format: uuid }
                      fulfillmentOrder:
                        type: object
                        required: [id, status, shippedQuantityGrams]
                        properties:
                          id: { type: string, format: uuid }
                          status: { type: string, enum: [pending_shipment, partially_shipped, shipped, completed] }
                          shippedQuantityGrams: { type: integer }
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '404': { $ref: '#/components/responses/Error404' }
        '409':
          description: SHIPMENT_DUPLICATE_TRACKING / QUANTITY_EXCEEDS_ALLOCATION / FULFILLMENT_NOT_SHIPPABLE
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/admin/v1/fulfillment/orders/{id}/receiver:
    post:
      operationId: updateFulfillmentReceiver
      security: [{ bearerAuth: [] }]
      summary: 发货前修改收货快照（version+1；审计；发货后 409 RECEIVER_LOCKED）
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receiverName, phone, province, city, district, detail]
              properties:
                receiverName: { type: string, maxLength: 20 }
                phone: { type: string }
                province: { type: string, maxLength: 20 }
                city: { type: string, maxLength: 20 }
                district: { type: string, maxLength: 20 }
                detail: { type: string, maxLength: 120 }
      responses:
        '200':
          description: 已更新（新版本）
          content:
            application/json:
              schema:
                type: object
                required: [data, requestId]
                properties:
                  requestId: { type: string }
                  data:
                    type: object
                    properties:
                      fulfillmentOrder:
                        type: object
                        properties:
                          id: { type: string, format: uuid }
                          receiverVersion: { type: integer }
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '404': { $ref: '#/components/responses/Error404' }
        '409':
          description: RECEIVER_LOCKED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/admin/v1/fulfillment/orders/{id}/complete:
    post:
      operationId: completeFulfillmentOrder
      security: [{ bearerAuth: [] }]
      summary: 管理员标记完成（仅 shipped；幂等；审计；order:manage）
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: 已完成（幂等返回现状）
          content:
            application/json:
              schema:
                type: object
                required: [data, requestId]
                properties:
                  requestId: { type: string }
                  data:
                    type: object
                    properties:
                      fulfillmentOrder:
                        type: object
                        properties:
                          id: { type: string, format: uuid }
                          status: { type: string, enum: [completed] }
                          completedBy: { type: string, enum: [admin] }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '404': { $ref: '#/components/responses/Error404' }
        '409':
          description: FULFILLMENT_NOT_SHIPPABLE（未全部发货）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /api/mini/v1/cards/{kind}/{id}:
    get:
      operationId: getMiniCard
      security: [{ bearerAuth: [] }]
      summary: 卡片只读投影（product 需上架；order/group/refund 仅本人）
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string, enum: [product, order, group, refund] } }
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 卡片摘要（按 kind 结构不同）
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
...
  /api/mini/v1/cards/options/{kind}:
    get:
      operationId: listMiniCardOptions
      summary: "授权业务卡片候选，最多最近50条"
      security: [{ bearerAuth: [] }]
      parameters: [{"name":"kind","in":"path","required":true,"schema":{"type":"string","enum":["product","order","group","refund"]}}]
      responses:
        '200': {"description":"授权业务卡片候选，最多最近50条","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"object","properties":{"items":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"label":{"type":"string"}}}}}},"requestId":{"type":"string"}}}}}}
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
...
  /api/admin/v1/cs/conversations/{id}/cards/{kind}/{ref}:
    get:
      operationId: getCsBusinessCard
      summary: "负责客服读取接待用户授权的卡片投影"
      security: [{ bearerAuth: [] }]
      parameters: [{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"kind","in":"path","required":true,"schema":{"type":"string","enum":["product","order","group","refund"]}},{"name":"ref","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}]
      responses:
        '200': {"description":"负责客服读取接待用户授权的卡片投影","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"object","properties":{}},"requestId":{"type":"string"}}}}}}
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
...
  /api/admin/v1/fulfillment/blocks:
    get:
      operationId: listFulfillmentBlocks
      summary: "order:manage读取暂停履约异常组及订单引用，最多100条"
      security: [{ bearerAuth: [] }]
      responses:
        '200': {"description":"order:manage读取暂停履约异常组及订单引用，最多100条","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"object","properties":{"items":{"type":"array","items":{"type":"object","properties":{"groupId":{"type":"string","format":"uuid"},"reason":{"type":"string","enum":["ZERO_ALLOCATION","INVALID_QUANTITY"]},"createdAt":{"type":"string","format":"date-time"},"orderIds":{"type":"array","items":{"type":"string","format":"uuid"}}}}}}},"requestId":{"type":"string"}}}}}}
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 管理员会话 token（admin 域）
    miniUserAuth:
      type: http
      scheme: bearer
      description: 小程序用户会话 token（user 域，与 admin 不可互用）
  responses:
...
    Error400:
      description: 参数校验失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Error401:
      description: 未认证
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Error404:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Error403:
      description: 无权限
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        details:
          type: array
          items: { type: string }
...
    AdminFulfillmentDetail:
      type: object
      required: [fulfillmentOrderId, orderId, orderNo, nickname, units, allocatedQuantityGrams, allocatedQuantityText, status, receiver, shipments]
      properties:
        fulfillmentOrderId: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        orderNo: { type: string }
        nickname: { type: string }
        units: { type: integer }
        allocatedQuantityGrams: { type: integer, description: 最小履约单位数（重量=克；计数=件） }
        allocatedQuantityText: { type: string, description: 原快照单位展示文本 }
        unit: { type: string, description: 原快照计量单位 }
        status: { type: string, enum: [pending_shipment, partially_shipped, shipped, completed] }
        receiver:
          type: object
          required: [name, phoneMasked, province, city, district, detail, version]
          properties:
            name: { type: string }
            phoneMasked: { type: string, description: 列表脱敏；导出 CSV 为明文且写审计 }
            province: { type: string }
            city: { type: string }
            district: { type: string }
            detail: { type: string }
            version: { type: integer, description: 收货快照版本（管理员修改递增） }
        shipments:
          type: array
          items:
            type: object
            required: [id, quantityGrams, quantityText, isReissue, company, trackingNo, shippedAt]
            properties:
              id: { type: string, format: uuid }
              quantityGrams: { type: integer }
              quantityText: { type: string }
              isReissue: { type: boolean }
              reissueReason: { type: [string, 'null'] }
              company: { type: string }
              trackingNo: { type: string }
              shippedAt: { type: string, format: date-time }
...